+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+// для состояния "Результат квиза"
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
