+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,9 @@
+import Foundation
+import UIKit
+
+// для состояния "Вопрос задан"
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
